# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EDCS API",
        "description": "The Elite Dangerous Companion Suite API provides galaxy data from Elite Dangerous to various consumers, including the EDCS frontend at https://edcs.app. Data is sourced from EDSM (Elite Dangerous Star Map), EDDN (Elite Dangerous Data Network), and the Frontier Companion API (CAPI).",
        "contact": {
            "name": "EDCS",
            "url": "https://edcs.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.edcs.app/api",
            "description": "EDCS API Server"
        }
    ],
    "paths": {
...
        "/frontier/capi/profile": {
            "get": {
                "tags": [
                    "Frontier CAPI"
                ],
                "summary": "Get the authenticated commander profile from Frontier CAPI",
                "description": "Fetches the live commander profile from the Frontier Companion API and confirms/updates the stored commander record.",
                "operationId": "04a52e1fac36e9d0cbbb5e147694d94d",
                "responses": {
                    "200": {
                        "description": "Commander profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Commander profile from Frontier CAPI",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/frontier/capi/journal": {
            "get": {
                "tags": [
                    "Frontier CAPI"
                ],
                "summary": "Get the commander journal for a given date",
                "description": "Get all journal entries.",
                "operationId": "f175382b7a5eb8385f17cd15543e974f",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Journal year",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2024
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Journal month (1–12)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "day",
                        "in": "query",
                        "description": "Journal day",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Journal entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/galnet/news": {
            "get": {
                "tags": [
                    "GalNet"
                ],
                "summary": "List GalNet news articles",
                "description": "Returns a paginated list of in-game GalNet news articles, most recent first.",
                "operationId": "ee0e4927700fcdafa7f4a21732584b45",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of GalNet articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GalnetNews"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/galnet/news/{slug}": {
            "get": {
                "tags": [
                    "GalNet"
                ],
                "summary": "Get a single GalNet news article by slug",
                "description": "Display a galnet news article.",
                "operationId": "9b58aec38ff2a99f330ad91415ceca16",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Article slug in format {date}-{title-kebab}, e.g. 16-aug-3310-the-assault-on-thor",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "16-aug-3310-the-assault-on-thor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GalNet article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GalnetNews"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
        "/galnet/news/{id}": {
            "delete": {
                "tags": [
                    "GalNet"
                ],
                "summary": "Delete a GalNet news article",
                "operationId": "91e8e157fbadd33418c648f02085891d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/stations/{slug}": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get a single station by slug",
                "description": "Retrieves a station, outpost, or megaship by its slug ({market_id}-{name}). Pass withSystem to embed the parent system and its bodies.",
                "operationId": "f0c8f6cdab8301614d6eba48f070c31e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Station slug in format {market_id}-{name}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "128016384-daedalus"
                        }
                    },
                    {
                        "name": "withSystem",
                        "in": "query",
                        "description": "Embed the parent system and its bodies",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Station",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Station"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GalnetNews": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "title": {
                        "type": "string",
                        "example": "The Assault on Thor"
                    },
                    "content": {
                        "type": "string",
                        "example": "<p>In a bold manoeuvre...</p>"
                    },
                    "audio_file": {
                        "type": "string",
                        "example": "https://audio.example.com/file.mp3",
                        "nullable": true
                    },
                    "uploaded_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "banner_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "16-aug-3310-the-assault-on-thor"
                    }
                },
                "type": "object"
            },
...
            "Station": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Daedalus"
                    },
                    "type": {
                        "type": "string",
                        "example": "Orbis Starport"
                    },
                    "body": {
                        "type": "string",
                        "example": "Sol",
                        "nullable": true
                    },
                    "distance_to_arrival": {
                        "type": "number",
                        "format": "float",
                        "example": 508
                    },
                    "controlling_faction": {
                        "type": "string",
                        "example": "Sol Constitution Party",
                        "nullable": true
                    },
                    "allegiance": {
                        "type": "string",
                        "example": "Federation",
                        "nullable": true
                    },
                    "government": {
                        "type": "string",
                        "example": "Democracy",
                        "nullable": true
                    },
                    "economy": {
                        "type": "string",
                        "example": "Industrial",
                        "nullable": true
                    },
                    "second_economy": {
                        "type": "string",
                        "example": "Refinery",
                        "nullable": true
                    },
                    "has_market": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_shipyard": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_outfitting": {
                        "type": "boolean",
                        "example": true
                    },
                    "other_services": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Restock",
                            "Repair",
                            "Contacts"
                        ]
                    },
                    "last_updated": {
                        "properties": {
                            "information": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "market": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "shipyard": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "outfitting": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "slug": {
                        "type": "string",
                        "example": "128016384-daedalus"
                    },
                    "system": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/System"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SystemInformation": {
                "properties": {
                    "allegiance": {
                        "type": "string",
                        "example": "Federation"
                    },
                    "government": {
                        "type": "string",
                        "example": "Democracy"
                    },
                    "population": {
                        "type": "integer",
                        "example": 22780000000
                    },
                    "security": {
                        "type": "string",
                        "example": "High"
                    },
                    "economy": {
                        "type": "string",
                        "example": "Refinery"
                    },
                    "controlling_faction": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Mother Gaia"
                            },
                            "state": {
                                "type": "string",
                                "example": "None"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SystemCoords": {
                "properties": {
                    "x": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "y": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "z": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "System": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id64": {
                        "type": "integer",
                        "example": 10477373803
                    },
                    "name": {
                        "type": "string",
                        "example": "Sol"
                    },
                    "coords": {
                        "$ref": "#/components/schemas/SystemCoords"
                    },
                    "slug": {
                        "type": "string",
                        "example": "10477373803-sol"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "information": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SystemInformation"
                            }
                        ],
                        "nullable": true
                    },
                    "bodies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SystemBody"
                        },
                        "nullable": true
                    },
                    "stations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Station"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum bearer token. Obtain via POST /auth/login or the Frontier SSO flow.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}